# src/main/resources/docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample Petstore server.  You can find \nout more about Swagger at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n",
    "version" : "1.0.0",
    "title" : "Time Dominator",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:9001",
  "basePath" : "/api-time-dominator/1.0.0",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/time-entries" : {
      "get" : {
        "tags" : [ "time-entry" ],
        "summary" : "Find time entries",
        "description" : "Find time entries",
        "operationId" : "findTimeEntries",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "dateFrom",
          "in" : "query",
          "description" : "Query param",
          "required" : false,
          "type" : "string",
          "format" : "date-time"
        }, {
          "name" : "dateTo",
          "in" : "query",
          "description" : "Query param",
          "required" : false,
          "type" : "string",
          "format" : "date-time"
        }, {
          "name" : "type",
          "in" : "query",
          "description" : "Query param",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "integer",
                "format" : "int32"
              }
            }
          }
        }
      }
    },
    "/time-entries/daily" : {
      "get" : {
        "tags" : [ "time-entry" ],
        "summary" : "Find daily time entries",
        "description" : "Find all types of daily time entries, including empty entries",
        "operationId" : "findDailyTimeEntries",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "date",
          "in" : "query",
          "description" : "Query param",
          "required" : true,
          "type" : "string",
          "format" : "date-time"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "integer",
                "format" : "int32"
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "petstore_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "http://petstore.swagger.io/oauth/dialog",
      "flow" : "implicit",
      "scopes" : {
        "write:pets" : "modify pets in your account",
        "read:pets" : "read your pets"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
...
}